change: moved the content of TimeStamp class into LongFormatLine class
