+++ 08.ls_object/lib/sections/long_formats.rb
@@ -4,4 +4,3 @@ require_relative 'long_formats/entire_file_status_width'
 require_relative 'long_formats/file_mode'
 require_relative 'long_formats/long_format_line'
 require_relative 'long_formats/permission'
-require_relative 'long_formats/time_stamp'
